Use custom irc-level keep-alive to reconnect automatically
ESTABLISHED TCP sockets will not close themselves if there's no traffic.

If lose (not local, but an intermediate router) connection, then it
comes back, the server might have forgotten about us, so if we don't
poke regularly the connection, we won't know if we're still online.

This could also be done with linux tcp(7) keepalive (TCP_KEEP*), but is
less portable.

We also workaround a "race" between disconnected state and keep alive
sending.